+++ src/lib/header/Header.svelte
@@ -58,6 +58,7 @@
 
     const navPages = [
         {path: "/", text: "Home"},
+		{path: "/mmt-2022", text: "MMT 2022"},
         {path: "/our-team", text: "Our Team"},
 		{path: "/sponsors", text: "Sponsors"},
         {path: "/past-exams", text: "Past Exams"},
